etc/Test/requirements.txt
pyserial
python-snap7